requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Django==4.0.6
djangorestframework==3.13.1
idna==3.3
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.11
